+++ src/ppom-middleware.test.ts
@@ -1,30 +1,6 @@
 import { buildFetchSpy, buildPPOMController } from '../test/test-utils';
 import { createPPOMMiddleware } from './ppom-middleware';
 
-Object.defineProperty(globalThis, 'fetch', {
-  writable: true,
-  value: () => undefined,
-});
-
-jest.mock('./ppom.ts', () => ({
-  PPOM: class PPOMClass {
-    #jsonRpcRequest;
-
-    constructor(jsonRpcRequest: any) {
-      this.#jsonRpcRequest = jsonRpcRequest;
-    }
-
-    validateJsonRpc = async () => {
-      return Promise.resolve();
-    };
-
-    free = () => undefined;
-
-    testJsonRPCRequest = () => this.#jsonRpcRequest();
-  },
-  ppomInit: () => undefined,
-}));
-
 describe('PPOMMiddleware', () => {
   it('should return PPOM Middleware when createPPOMMiddleware function called', () => {
     buildFetchSpy();